+++ module/campaign/runtime/spec.go
@@ -26,7 +26,7 @@ func OpenAPISpec() *openapi3.T {
 
 	return &openapi3.T{
 		OpenAPI: "3.0.3",
-		Info: &openapi3.Info{Title: "Campaign API", Version: "2.5.4"},
+		Info: &openapi3.Info{Title: "Campaign API", Version: "2.5.5"},
 		Paths: openapi3.NewPaths(
 			openapi3.WithPath("/campaigns", &openapi3.PathItem{Post: createOperation(), Get: listOperation()}),
 			openapi3.WithPath("/campaigns/{id}", &openapi3.PathItem{Get: getOperation(), Patch: updateOperation(), Delete: deleteOperation()}),